# Summary
- to close #19505

- saving 40% API service launch time (4627.76 ms -> 2812.27 ms on MacOS M4 Max), by deferring imports of tracing classes and SDK, which also helps to skip unnecessary network requests on starting introduced by Litellm from opik.

- Before this PR:
  - `ext_import_modules` process takes 3406.33 ms
  ```
  ...
  ^[[A2025-05-11 15:43:27,216 INFO [_client.py:1038]  HTTP Request: GET https://raw.githubusercontent.com/BerriAI/litellm/main/model_prices_and_context_window.json "HTTP/1.1 200 OK"
  2025-05-11 15:43:28,581 INFO [app_factory.py:99]  Loaded ext_import_modules (3406.33 ms)
  ...
  2025-05-11 15:43:29,115 INFO [app_factory.py:35]  Finished create_app (4627.76 ms)
  ```

- With this PR:
  - `ext_import_modules` process takes 1548.72 ms
  ```
  ...
  2025-05-11 15:53:33,174 INFO [app_factory.py:99]  Loaded ext_import_modules (1548.72 ms)
  ...
  2025-05-11 15:53:33,703 INFO [app_factory.py:35]  Finished create_app (2812.27 ms)
  ```

# Screenshots

| Before | After |
|--------|-------|
| ...    | ...   |

# Checklist

> [!IMPORTANT]  
> Please review the checklist below before submitting your pull request.

- [ ] This change requires a documentation update, included: [Dify Document](https://github.com/langgenius/dify-docs)
- [x] I understand that this PR may be closed in case there was no previous discussion or issues. (This doesn't apply to typos!)
- [x] I've added a test for each change that was introduced, and I tried as much as possible to make a single atomic change.
- [x] I've updated the documentation accordingly.
- [x] I ran `dev/reformat`(backend) and `cd web && npx lint-staged`(frontend) to appease the lint gods